update
更新poco 版本号3fc3e5f

+++ poco/Net/src/ICMPPacketImpl.cpp
@@ -81,13 +81,13 @@ unsigned short ICMPPacketImpl::checksum(UInt16 *addr, Int32 len)
 	UInt16 answer;
 	Int32 sum = 0;
 
-	while (nleft > 1)
+	while (nleft > 1)  
 	{
 		sum   += *w++;
 		nleft -= sizeof(UInt16);
 	}
 
-	if (nleft == 1)
+	if (nleft == 1) 
 	{
 		UInt16 u = 0;
 		*(UInt8*) (&u) = *(UInt8*) w;
@@ -96,7 +96,7 @@ unsigned short ICMPPacketImpl::checksum(UInt16 *addr, Int32 len)
 
 	sum = (sum >> 16) + (sum & 0xffff);
 	sum += (sum >> 16);
-	answer = static_cast<UInt16>(~sum);
+	answer = ~sum;
 	return answer;
 }
